# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Livestock Price Board API
  version: 0.1.0-phase1
  summary: Phase 1 contract. Auth, farm and herd records, price board, admin verification and reference prices.
  description: |
    Stack-agnostic contract for the Phase 1 build (weeks 3 to 8). NestJS or Laravel
    can implement it unchanged. Phase 2 (listings, offers, deals) and Phase 3
    (shipments) will be added as separate tag groups without breaking these paths.

    Conventions
    - All responses are JSON. Errors use RFC 9457 `application/problem+json`.
    - Every mutating request from a mobile client carries `Idempotency-Key` so an
      offline queue can retry safely. Same key + same body returns the original
      response; same key + different body returns 422.
    - Lists are cursor paginated: `?cursor=&limit=` in, `next_cursor` out.
    - Money is a decimal string in Philippine pesos, e.g. "185.50". Never a float.
    - Weights are decimal strings in kilograms.
    - Timestamps are RFC 3339 UTC. Dates are `YYYY-MM-DD`.
    - Locations are PSGC codes (text). The client never sends free-text place names.
    - Every price returned carries `source`, `location_code` and `sample_count`.
      A client must display all three. A bare number is a defect.
    - Records that can be edited offline (farms, lots) carry `version`. PATCH must
      send `If-Match: <version>`; a stale version returns 409 with the server copy.

    Public board
...
      operations; nothing else changes.
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /deals:
    get:
      tags: [Marketplace]
      operationId: listMyDeals
      summary: My deals as farmer or buyer
      parameters:
        - name: state
          in: query
          schema: { $ref: '#/components/schemas/DealState' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Deal' }
                  next_cursor: { type: [string, 'null'] }
...
  /admin/disputes:
    get:
      tags: [Admin]
      operationId: adminListDisputes
      summary: Disputes, open first
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [open, under_review, resolved_settled, resolved_refunded, dismissed] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Dispute' }
  /admin/disputes/{dispute_id}/resolve:
    post:
      tags: [Admin]
      operationId: adminResolveDispute
      summary: Resolve a dispute
      description: settled keeps the deal and lets its price join the board (optionally at a corrected weight); refunded excludes it; dismissed returns the deal to delivered.
      parameters:
        - name: dispute_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome, resolution]
              properties:
                outcome: { type: string, enum: [settled, refunded, dismissed] }
                resolution: { type: string, minLength: 3, maxLength: 1000 }
                delivered_weight_kg: { $ref: '#/components/schemas/Weight' }
                deposit:
                  type: string
                  enum: [release_to_farmer, refund_to_buyer, hold]
                  description: Where a paid deposit goes. Defaults follow the outcome (settled releases, refunded refunds, dismissed holds).
      responses:
        '200':
          description: Resolved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dispute' }
  /admin/deals:
    get:
      tags: [Admin]
      operationId: adminListDeals
      summary: Deals across all users, newest first
      parameters:
        - name: state
          in: query
          schema: { $ref: '#/components/schemas/DealState' }
        - name: species
          in: query
          schema: { $ref: '#/components/schemas/Species' }
        - name: province_code
          in: query
          schema: { $ref: '#/components/schemas/PsgcCode' }
        - name: outliers_only
          in: query
          schema: { type: boolean, default: false }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Deal' }
                  next_cursor: { type: [string, 'null'] }
  /admin/deals/{deal_id}/outlier-review:
    post:
      tags: [Admin]
      operationId: adminReviewOutlier
      summary: Let a flagged settlement count, or keep it out
      parameters:
        - name: deal_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [counts_for_price]
              properties:
                counts_for_price: { type: boolean }
                note: { type: [string, 'null'], maxLength: 500 }
      responses:
        '200':
          description: Reviewed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deal' }
  # ------------------------------------------------------------------ Logistics
...
  /admin/verification-queue:
    get:
      tags: [Admin]
      operationId: adminVerificationQueue
      summary: Users with pending documents, oldest first
      parameters:
        - name: role
          in: query
          schema: { $ref: '#/components/schemas/Role' }
        - name: province_code
          in: query
          schema: { $ref: '#/components/schemas/PsgcCode' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/VerificationCase' }
                  next_cursor: { type: [string, 'null'] }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/users/{user_id}:
    get:
      tags: [Admin]
      operationId: adminGetUser
      summary: User detail with documents, farms and audit history
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationCase' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/users/{user_id}/verification:
    post:
      tags: [Admin]
      operationId: adminSetVerification
      summary: Verify, reject or suspend a user
      description: Writes an `admin_audit_log` row with before and after. Sends the user an SMS on any change.
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [verified, rejected, suspended]
                notes: { type: string, maxLength: 1000, description: Shown to the user when rejected. }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/documents/{document_id}/review:
    post:
      tags: [Admin]
      operationId: adminReviewDocument
      summary: Accept or reject one document
      parameters:
        - name: document_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [verified, rejected] }
                notes: { type: string, maxLength: 1000 }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserDocument' }
  /admin/documents/{document_id}/file:
    get:
      tags: [Admin]
      operationId: adminGetDocumentFile
      summary: Short-lived signed URL to view a document
      description: URL expires in 5 minutes. Access is logged.
      parameters:
        - name: document_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string, format: uri }
                  expires_at: { type: string, format: date-time }
  /admin/reference-prices:
    get:
      tags: [Admin]
      operationId: adminListReferencePrices
      summary: Current reference price per province, species and class
      parameters:
        - name: province_code
          in: query
          schema: { $ref: '#/components/schemas/PsgcCode' }
        - name: species
          in: query
          schema: { $ref: '#/components/schemas/Species' }
        - name: include_history
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ReferencePrice' }
    post:
      tags: [Admin]
      operationId: adminSetReferencePrice
      summary: Set a reference price
      description: |
        Inserts a new row; history is never overwritten. `effective_from` may not be
        earlier than 7 days ago. A price more than 30 percent from the previous row
        for the same key is accepted but flagged `large_change: true` in the response
        and in the audit log, so a typo is caught by the second admin.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReferencePriceInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReferencePrice' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /admin/reference-prices/import:
    post:
      tags: [Admin]
      operationId: adminImportReferencePrices
      summary: Bulk import from a CSV (PSA farmgate series)
      description: |
        CSV columns: province_code, species, weight_class_label (optional), unit, price,
        effective_from, source. Validated as a whole; any bad row rejects the file with
        line numbers. Nothing is written on a partial failure.
      requestBody:
        required: true
        content:
          text/csv:
            schema: { type: string }
      responses:
        '201':
          description: Imported
          content:
            application/json:
              schema:
                type: object
                properties:
                  inserted: { type: integer }
                  large_changes:
                    type: array
                    items: { $ref: '#/components/schemas/ReferencePrice' }
        '422':
          description: Validation errors with line numbers.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /admin/restricted-zones:
    get:
      tags: [Admin]
      operationId: adminListRestrictedZones
      summary: Active and scheduled movement restrictions
      parameters:
        - name: active_on
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RestrictedZone' }
    post:
      tags: [Admin]
      operationId: adminCreateRestrictedZone
      summary: Mark a location restricted for a species
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RestrictedZoneInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RestrictedZone' }
  /admin/restricted-zones/{zone_id}:
    patch:
      tags: [Admin]
      operationId: adminEndRestrictedZone
      summary: Set or change the end date
      parameters:
        - name: zone_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ends_on]
              properties:
                ends_on: { type: string, format: date }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RestrictedZone' }
  /admin/price-snapshots/refresh:
    post:
      tags: [Admin]
      operationId: adminRefreshSnapshots
      summary: Run the snapshot job for a date now
      description: Calls `refresh_price_snapshots(date)`. Normally the nightly scheduler does this. Audited.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                date: { type: string, format: date }
                window_days: { type: integer, default: 7 }
      responses:
        '200':
          description: Rows written.
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows: { type: integer }
  /admin/audit-log:
    get:
      tags: [Admin]
      operationId: adminAuditLog
      summary: Admin actions, newest first
      parameters:
        - name: admin_id
          in: query
          schema: { type: string, format: uuid }
        - name: action
          in: query
          schema: { type: string }
        - name: target_id
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AuditEntry' }
                  next_cursor: { type: [string, 'null'] }
  /admin/users:
    get:
      tags: [Admin]
      operationId: adminListUsers
      summary: Everyone, filterable
      parameters:
        - name: role
          in: query
          schema: { $ref: '#/components/schemas/Role' }
        - name: verification
          in: query
          schema: { $ref: '#/components/schemas/VerificationStatus' }
        - name: q
          in: query
          description: Name or phone fragment.
          schema: { type: string, maxLength: 60 }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AdminUserRow' }
                  next_cursor: { type: [string, 'null'] }
  /admin/reports/summary:
    get:
      tags: [Admin]
      operationId: adminReportSummary
      summary: Health of the whole system for a period
      description: Volume and price by species, users by role, time to settle, dispute rate, deposit outcomes, hauled share, and municipalities with listings but too few settled deals for a live median.
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
          description: Default 30 days ago.
        - name: to
          in: query
          schema: { type: string, format: date }
          description: Default today.
        - name: province_code
          in: query
          schema: { $ref: '#/components/schemas/PsgcCode' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportSummary' }
  /admin/reports/deals.csv:
    get:
      tags: [Admin]
      operationId: adminReportDealsCsv
      summary: Settled and other deals in the period as CSV
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: province_code
          in: query
          schema: { $ref: '#/components/schemas/PsgcCode' }
        - name: state
          in: query
          schema: { $ref: '#/components/schemas/DealState' }
      responses:
        '200':
          description: One row per deal.
          content:
            text/csv:
              schema: { type: string }
  # ------------------------------------------------------------------ Admin auth
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token, 15 minute lifetime. Claims: `sub` (user id), `roles` (array),
        `ver` (verification status), `jti`. Role checks happen server-side on every
        request; the client uses the claims only to choose which screens to show.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated UUID. Stored 48 hours.
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
  responses:
...
    Forbidden:
      description: Role or ownership check failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Not found, or not visible to this user.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unprocessable:
      description: Validation failed. `errors` lists field problems.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    PsgcCode:
      type: string
      pattern: '^[0-9]{9,10}$'
      description: Philippine Standard Geographic Code.
      example: '0304924000'
...
    Weight:
      type: string
      pattern: '^[0-9]+(\.[0-9]{1,2})?$'
      description: Kilograms as a decimal string.
      example: '92.5'
    Role:
      type: string
      enum: [farmer, buyer, hauler, admin]
    VerificationStatus:
      type: string
      enum: [pending, verified, rejected, suspended]
    Species:
      type: string
      enum: [hog, cattle, carabao, goat, native_chicken]
...
    User:
      type: object
      required: [id, phone_masked, full_name, preferred_lang, roles, verification, created_at]
      properties:
        id: { type: string, format: uuid }
        phone_masked: { type: string, example: '+63917*****67' }
        full_name: { type: string }
        preferred_lang: { $ref: '#/components/schemas/LangCode' }
        roles:
          type: array
          items: { $ref: '#/components/schemas/Role' }
        verification: { $ref: '#/components/schemas/VerificationStatus' }
        verified_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
    UserDocument:
      type: object
      required: [id, doc_type, status, uploaded_at]
      properties:
        id: { type: string, format: uuid }
        doc_type: { $ref: '#/components/schemas/DocType' }
        status: { $ref: '#/components/schemas/VerificationStatus' }
        notes: { type: [string, 'null'], description: 'Reviewer note, shown to the user when rejected.' }
        uploaded_at: { type: string, format: date-time }
        reviewed_at: { type: [string, 'null'], format: date-time }
...
    VerificationCase:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        documents:
          type: array
          items: { $ref: '#/components/schemas/UserDocument' }
        farms:
          type: array
          items: { $ref: '#/components/schemas/Farm' }
        oldest_pending_at: { type: [string, 'null'], format: date-time }
...
    DealState:
      type: string
      enum: [accepted, hauler_assigned, in_transit, delivered, settled, cancelled, disputed, refunded]
...
    Deal:
      type: object
      required: [id, listing_id, offer_id, farmer_id, buyer_id, species, unit, state, agreed_price, agreed_heads, needs_hauler, accepted_at]
      properties:
        id: { type: string, format: uuid }
        listing_id: { type: string, format: uuid }
        offer_id: { type: string, format: uuid }
        farmer_id: { type: string, format: uuid }
        farmer_name: { type: string }
        buyer_id: { type: string, format: uuid }
        buyer_name: { type: string }
        species: { $ref: '#/components/schemas/Species' }
        weight_class: { $ref: '#/components/schemas/WeightClass' }
        unit: { $ref: '#/components/schemas/PriceUnit' }
        state: { $ref: '#/components/schemas/DealState' }
        agreed_price: { $ref: '#/components/schemas/Money' }
        agreed_heads: { type: integer }
        agreed_weight_kg: { $ref: '#/components/schemas/MoneyNullable' }
        estimated_total: { $ref: '#/components/schemas/MoneyNullable' }
        delivered_heads: { type: [integer, 'null'] }
        delivered_weight_kg: { $ref: '#/components/schemas/MoneyNullable' }
        final_total: { $ref: '#/components/schemas/MoneyNullable', description: delivered weight times agreed price, or heads times price for per-head units. }
        needs_hauler: { type: boolean }
        payment_method: { type: [string, 'null'] }
        payment_reference: { type: [string, 'null'] }
        buyer_paid_at: { type: [string, 'null'], format: date-time }
        farmer_confirmed_at: { type: [string, 'null'], format: date-time }
        outlier_flag: { type: boolean }
        counts_for_price: { type: boolean }
        location: { $ref: '#/components/schemas/LocationWithPath' }
        dropoff: { $ref: '#/components/schemas/LocationWithPath' }
        shipment: { $ref: '#/components/schemas/ShipmentSummary' }
        deposit_required: { type: boolean }
        deposit: { $ref: '#/components/schemas/DepositSummary' }
        events:
          type: array
          items: { $ref: '#/components/schemas/DealEvent' }
        accepted_at: { type: string, format: date-time }
        delivered_at: { type: [string, 'null'], format: date-time }
        settled_at: { type: [string, 'null'], format: date-time }
        cancelled_at: { type: [string, 'null'], format: date-time }
    Dispute:
      type: object
      required: [id, deal_id, raised_by, reason, status, created_at]
      properties:
        id: { type: string, format: uuid }
        deal_id: { type: string, format: uuid }
        raised_by: { type: string, format: uuid }
        raised_by_role: { type: string, enum: [farmer, buyer] }
        reason: { type: string }
        details: { type: [string, 'null'] }
        status: { type: string, enum: [open, under_review, resolved_settled, resolved_refunded, dismissed] }
        resolution: { type: [string, 'null'] }
        resolved_by: { type: [string, 'null'], format: uuid }
        deal: { $ref: '#/components/schemas/Deal' }
        created_at: { type: string, format: date-time }
        resolved_at: { type: [string, 'null'], format: date-time }
...